Switch DEPO for GPU for using Eco class
The change unifies the usage of DEPO related methods for both CPU and GPU.
As of this commit the GpuEco class will no longer be improved and eventually
deprecated and removed.
Only DEPO is using Eco class method right now - StEP and other tools need to
be updated and switched from GpuEco to Eco in the next incoming changes.
This commit also starts the process of refactoring Trigger of tuning phase.
The Trigger class will be updated soon.