Fix STEP (logging and minimal cap)
1. Logging results from step to result.csv stores averaged result of N runs
and each of averaged results commented with #. The powercap value logged for
the commended results entry was fixed on max power cap value - now it is fixed
and stores actual power cap value.
2. Exploring various power caps in STEP with Linear Search pattern suffered from
missing the lowest power cap when the step was a divisor of the available power
caps range - now it is fixed and the lowest power cap will also be included in
energy profiling.